+++ blockers/youratorBlockers.ts
@@ -1,4 +1,4 @@
-import { type Blocker } from '../../../../schemas/blocker'
+import { type Blocker } from '../schemas/blocker'
 
 export const youratorBlockerOptions: Blocker[] = [
   {
