+++ schemas/activatorStrategy.ts
