+++ schemas/platformName.ts
