+++ schemas/selectorStrategy.ts
