+++ tsconfig.script.json
@@ -8,12 +8,10 @@
     "strictNullChecks": true
   },
   "include": [
-    ".eslintrc.cjs",
-    "vite.config.mts",
-    "vitest.config.ts",
-    "manifest.config.ts",
-    "tailwind.config.ts",
-    "postcss.config.cjs",
-    "commitlint.config.ts",
+    "package.json",
+    "scripts/**/*.js",
+    "scripts/**/*.ts",
+    "blockers/**/*.ts",
+    "schemas/**/*.ts",
   ]
 }
